pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "github-standards"
description = "Hooks for use in a DBT development environment"
readme = "README.md"
keywords = ["pre-commit", "dbt", "hooks", "security"]
requires-python = ">=3.13"
dependencies = [
    "presidio-analyzer>=2.2.360",
    "proxy-py>=2.4.10",
    "pyyaml>=6.0.3",
    "requests>=2.32.4",
    "en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "gitpython>=3.1.45",
]
version = "v1.1.2"

# Presence enables setuptools-scm
[tool.setuptools_scm]


[project.scripts]
hooks-cli = "src.hooks.cli:main"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]
namespaces = false

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.13.2",
    "pre-commit>=4.5.0",
    "pytest-cov>=5.0.0",
    "requests-mock>=1.12.1",
    "pytest-subprocess>=1.5.3",
    "setuptools-scm>=8",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 125
indent-width = 4
target-version = "py312"
extend-select = ["T201"]

[tool.coverage.run]
omit = ['tests/*']
